pyproject.toml
[project]
name = "stock-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "dotenv>=0.9.9",
    "google-genai>=1.52.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "plotly>=6.5.0",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "yfinance>=0.2.66",
]
